% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceGWAS_object.R
\docType{class}
\name{ceGWAS}
\alias{ceGWAS}
\title{The central object of cegwas2}
\format{\code{\link{R6Class}} object.}
\usage{
ceGWAS
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.
}
\description{
The central object of cegwas2
}
\section{Fields}{

\describe{
\item{\code{trait_name}}{name of trait, extracted from input data frame}

\item{\code{strains}}{name of strains provided in input data frame}

\item{\code{phenotype}}{phenotypes provided by the user}

\item{\code{processed_phenotype}}{output generated by \code{\link{process_phenotypes}}}

\item{\code{mapping}}{output generated by \code{\link{perform_mapping}}}

\item{\code{peak_intervals}}{identified confidence intervals}

\item{\code{fine_mapping}}{more details to come as we create methods
#'@section Methods:
\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/AndersenLab/cegwas2}
  \item{\code{new(phenotype)}}{This method is used to create object of this class with \code{phenotype},
  which is a two-column data frame with colnames strain and trait. Trait name can be user specified}

  \item{\code{set_markers(genotype_matrix)}}{This method sets genotype matrix to
  something other than the default provided by \code{cegwas2}. This will also generate a corresponding kinship matrix}
  \item{\code{run_mapping(...)}}{This method executes the \code{perform_mappng} function from \code{cegwas2}}
  }}
}}

\keyword{datasets}
