% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{cegwas2}
\alias{cegwas2}
\alias{cegwas2-package}
\title{\emph{C. elegans} genome-wide association studies 2}
\description{
A set of tools for studying natural variation in \emph{C. elegans}.
}
\section{Package options}{


Devtools uses the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{devtools.path}: path to use for \code{\link{dev_mode}}

  \item \code{devtools.name}: your name, used when signing draft
    emails.
}
}

