% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{cendr_dataset_release}
\alias{cendr_dataset_release}
\title{cendr_dataset_release}
\format{An object of class \code{numeric} of length 1.}
\usage{
cendr_dataset_release
}
\description{
The CeNDR dataset being used.
}
\keyword{datasets}
