% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db}
\alias{get_db}
\title{Get a dataframe from the CeNDR database}
\usage{
get_db(table = "wormbase_gene", renew = FALSE)
}
\arguments{
\item{table}{The table to request. [\strong{Default:} \code{"wormbase_gene"}]}

\item{renew}{Force download of the latest database. Then return the requested table. [\strong{Default:} \code{FALSE}]}
}
\value{
Dataframe with the requested table.
}
\description{
\code{get_db} downloads and caches the CeNDR database and fetches requested tables.
}
\details{
The database has five tables:

\itemize{
     \item \strong{homologs} - A table of homologs and orthologs.
     \item \strong{metadata} - data about the data in this database.
     \item \strong{strain} - A replicate of the WI Strain Database. Contains information on their names, isotypes, isolation location, and more.
     \item \strong{wormbase_gene} - Database of the gff-based gene models.
     \item \strong{wormbase_gene_summary} - A summary of the genes in \emph{C. elegans}.
}
}
\examples{
get_db()
}
