% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinship}
\alias{kinship}
\title{C. elegans relatedness matrix}
\format{A data frame with 249 rows, 249 variables:
\describe{
  \item{Strain1}{The relatedness of Strain1 to all other strains}
  \item{Strain2}{The relatedness of Strain2 to all other strains}
  ...
}}
\source{
\url{https://elegansvariation.org/}
}
\usage{
kinship
}
\description{
An N x N matrix containing the additive relatedness matrix in the C. elegans population
}
\keyword{datasets}
