% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_mappings.R
\name{perform_mapping}
\alias{perform_mapping}
\title{Perform GWA analysis}
\usage{
perform_mapping(phenotype = NULL, genotype = cegwas2::snps,
  kinship = cegwas2::kinship, P3D = FALSE, min.MAF = 0.05,
  map_by_chrom = FALSE)
}
\arguments{
\item{phenotype}{a data frame with columns: strain and phenotype [\strong{Default:} \code{"wormbase_gene"}]}

\item{genotype}{a genotype matrix in the format [\strong{Default:} \code{cegwas2::snps}]}

\item{kinship}{a N x N relatedness matrix in the format [\strong{Default:} \code{cegwas2::kinship}]}

\item{P3D}{TRUE/FALSE - TRUE refers to the EMMAx algorithm, FALSE refers to EMMA algorithm}

\item{map_by_chrom}{TRUE/FALSE - BLUP residual mappings from [\strong{Bloom, J. S. et al. 2015}]
[\strong{Default:} \code{FALSE}]}

\item{MAF}{a value ranging from 0 - 1 that determines the minimum minor allele frequencey
a marker must have to be used in association mapping [\strong{Default:} \code{0.05}]}
}
\value{
a dataframe with the following columns
\itemize{
     \item \strong{CHROM} - Chromosome name
     \item \strong{POS} - Physical position of marker
     \item \strong{marker} - Marker name
     \item \strong{trait} - Trait name of input phenotype
     \item \strong{BF} - Bonferroni-corrected p-value threshold
     \item \strin{log10p} - [\code{-log10}] transformation of the p-value for the indicated marker
}
}
\description{
\code{perform_mapping} runs association mapping on a phenotype
}
\examples{
get_db()
}
