% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snps}
\alias{snps}
\title{C. elegans SNV set}
\format{A data frame with 37851 rows, 4 variables, and 249 strains (as variables):
\describe{
  \item{CHROM}{Chromosome of SNV}
  \item{POS}{Physical position of SNV}
  \item{REF}{Reference allele}
  \item{ALT}{Alternate allele}
  \item{NIC511}{The first of 249 strains with genotypes}
  ...
}}
\source{
\url{https://elegansvariation.org/}
}
\usage{
snps
}
\description{
A data set that contains bi-allelic markers in the C. elegans population.
Values of -1 corresponds to the REF allele, 1 correspond to the ALT allele
}
\keyword{datasets}
