% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenotypes.R
\name{process_phenotypes}
\alias{process_phenotypes}
\title{Process phenotypes for mapping}
\usage{
process_phenotypes(df, summarize_replicates = "mean",
  prune_method = "BAMF", remove_outliers = TRUE, threshold = 2)
}
\arguments{
\item{df}{a dataframe with a strain column and columns for each trait for processing.}

\item{summarize_replicates}{summarization method, currently limited to "mean" or "median"}

\item{prune_method}{method for eliminating outliers, currently limited to "BAMF", "Z", "TUKEY", "MAD"}

\item{remove_outliers}{boolean to specify if outliers should be eliminated within the function.
If FALSE, additional columns will be output specifying if the strain phenotype is an outier.}

\item{threshold}{integer value defining the threshold for outlier removal, default = 2}
}
\value{
Output is a dataframe with
}
\description{
\code{process_phenotypes} takes input raw phenotype data and converts strain names to
isotype names, summarizes replicate data, and eliminates outliers.
}
