% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenotypes.R
\name{resolve_isotypes}
\alias{resolve_isotypes}
\title{Resolve strain names to isotypes}
\usage{
resolve_isotypes(...)
}
\arguments{
\item{strains}{a vector of strain names to change to isotype names.}
}
\value{
Output is a vector with isotype names
}
\description{
\code{resolve_isotypes} takes a vector of strain names and converts them to CeNDR-defined isotype names
}
