% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenotypes.R
\name{resolve_isotypes}
\alias{resolve_isotypes}
\title{Resolve strain names to isotypes}
\usage{
resolve_isotypes(strains2resolve, isotype_lookup = generate_isotype_lookup())
}
\arguments{
\item{strains2resolve}{a vector of strain names to change to isotype names.}

\item{isotype_lookup}{a tibble that contains two columns - strain, previous_name, isotype.
This is used to change input strain names. Default option queries CeNDR isotypes}
}
\value{
Output is a vector with isotype names
}
\description{
\code{resolve_isotypes} takes a vector of strain names and converts them to CeNDR-defined isotype names
}
